# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SocialPredict API
  version: 0.1.0
  description: >
    HTTP contract for the SocialPredict backend service. This spec is generated
    from the Go handlers and DTOs under backend/server and backend/handlers.
    Routes wrapped by the shared security middleware can return a shared
    transport-level `429 text/plain` response before reaching handler-specific
    JSON envelope logic when the rate limit is exceeded. Per-operation `429`
    responses document that middleware behavior where it applies.
servers:
  - url: http://localhost:8080
    description: Local dev server
...
paths:
...
  /v0/login:
    post:
      operationId: loginUser
      tags: [Auth]
      summary: Authenticate user
      description: >
        Validates username and password and returns a JWT bearer token for subsequent
        authenticated requests. Users flagged for password rotation still receive the
        normal token plus `mustChangePassword: true`; server-side enforcement limits
        what they can do with it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid JSON payload or validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '401':
          description: Invalid username or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidCredentials:
                  value:
                    ok: false
                    reason: INVALID_CREDENTIALS
        '500':
          description: Database access or token creation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          description: Login rate limit exceeded by middleware. Other protected routes can also return middleware-level 429 responses as described in the API description.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PlainTextErrorResponse'
...
  /v0/setup:
    get:
      operationId: getSetupConfiguration
      tags: [Config]
      summary: Get economics configuration
      description: Returns the current economic configuration used by the backend. This route is public in the current backend implementation.
      responses:
        '200':
          description: Economics configuration returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Economics'
        '500':
          description: Failed to load or encode configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/setup/frontend:
    get:
      operationId: getFrontendSetupConfiguration
      tags: [Config]
      summary: Get frontend configuration
      description: Returns minimal configuration values needed by the frontend, such as chart precision. This route is public in the current backend implementation.
      responses:
        '200':
          description: Frontend configuration returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendConfig'
        '500':
          description: Failed to load or encode configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets:
    get:
      operationId: listMarkets
      tags: [Markets]
      summary: List markets
      description: >
        Returns markets with optional filtering by status and creator. When status is
        provided, results are backed by the status-specific listing path used by the
        live handler.
      parameters:
        - in: query
          name: status
          required: false
          description: >
            Filter by market status. Accepted values are active, closed, resolved, and all
            (or omitted for all markets).
          schema:
            type: string
            enum: [active, closed, resolved, all]
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return. Malformed, non-positive, or out-of-range values fall back to the live default of 50.
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results. Malformed or negative values fall back to 0.
          schema:
            type: integer
        - in: query
          name: created_by
          required: false
          description: Only include markets created by the supplied username.
          schema:
            type: string
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '400':
          description: Invalid or unsupported status filter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
    post:
      operationId: createMarket
      tags: [Markets]
      summary: Create a market
      description: Creates a new prediction market for the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMarketRequest'
      responses:
        '201':
          description: Market created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketResponse'
        '400':
          description: Request decoding failed, validation failed, or the creator lacks funds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
                validationFailed:
                  value:
                    ok: false
                    reason: VALIDATION_FAILED
                insufficientBalance:
                  value:
                    ok: false
                    reason: INSUFFICIENT_BALANCE
        '401':
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                authenticationRequired:
                  value:
                    ok: false
                    reason: AUTHENTICATION_REQUIRED
        '403':
          description: Password change required before creating markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                passwordChangeRequired:
                  value:
                    ok: false
                    reason: PASSWORD_CHANGE_REQUIRED
        '404':
          description: Authenticated user record could not be loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                notFound:
                  value:
                    ok: false
                    reason: NOT_FOUND
        '500':
          description: Unexpected server error during market creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/search:
    get:
      operationId: searchMarkets
      tags: [Markets]
      summary: Search markets
      description: >
        Searches markets by question title with optional status filtering. Supports both
        "query" and legacy "q" parameters.
      parameters:
        - in: query
          name: query
          required: false
          description: Search query (preferred parameter name).
          schema:
            type: string
        - in: query
          name: q
          required: false
          description: Legacy search query parameter (used if "query" is not provided).
          schema:
            type: string
        - in: query
          name: status
          required: false
          description: Filter by market status; same accepted values as /v0/markets.
          schema:
            type: string
            enum: [active, closed, resolved, all]
        - in: query
          name: limit
          required: false
          description: Maximum number of primary results to return. Malformed, non-positive, or out-of-range values fall back to the live default of 20.
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          description: Number of primary results to skip before collecting. Malformed or negative values fall back to 0.
          schema:
            type: integer
      responses:
        '200':
          description: Search completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Missing query parameter or invalid status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '500':
          description: Unexpected server error during search.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/status:
    get:
      operationId: listAllMarketStatuses
      tags: [Markets]
      summary: List markets across all statuses
      description: Convenience endpoint equivalent to `/v0/markets?status=all`.
      parameters:
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return. Malformed, non-positive, or out-of-range values fall back to the live default of 100.
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results. Malformed or negative values fall back to 0.
          schema:
            type: integer
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/status/{status}:
    get:
      operationId: listMarketsByStatus
      tags: [Markets]
      summary: List markets by status
      description: Returns markets filtered by a specific lifecycle status.
      parameters:
        - in: path
          name: status
          required: true
          description: >
            Market status to filter by. Accepted values are active, closed, resolved, and all.
          schema:
            type: string
            enum: [active, closed, resolved, all]
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return. Malformed, non-positive, or out-of-range values fall back to the live default of 100.
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results. Malformed or negative values fall back to 0.
          schema:
            type: integer
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '400':
          description: Invalid or unsupported status value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/active:
    get:
      operationId: listLegacyActiveMarkets
      tags: [Markets]
      summary: Legacy active markets route
      description: Backward-compatible alias for `/v0/markets?status=active`.
      responses:
        '200':
          description: Active markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '500':
          description: Unexpected server error while listing active markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/closed:
    get:
      operationId: listLegacyClosedMarkets
      tags: [Markets]
      summary: Legacy closed markets route
      description: Backward-compatible alias for `/v0/markets?status=closed`.
      responses:
        '200':
          description: Closed markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '500':
          description: Unexpected server error while listing closed markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/resolved:
    get:
      operationId: listLegacyResolvedMarkets
      tags: [Markets]
      summary: Legacy resolved markets route
      description: Backward-compatible alias for `/v0/markets?status=resolved`.
      responses:
        '200':
          description: Resolved markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '500':
          description: Unexpected server error while listing resolved markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/{id}:
    get:
      operationId: getMarketDetails
      tags: [Markets]
      summary: Get market details
      description: >
        Fetches a single market by its identifier, including probability history and
        aggregate statistics.
      parameters:
        - in: path
          name: id
          required: true
          description: Numeric identifier of the market.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Market returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketDetailsResponse'
        '400':
          description: Invalid market ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '404':
          description: Market not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                notFound:
                  value:
                    ok: false
                    reason: NOT_FOUND
        '500':
          description: Unexpected server error while fetching market details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
...
  /v0/markets/bets/{marketId}:
    get:
      tags: [Bets, Markets]
      operationId: listMarketBets
      summary: List bets for a market
      description: Returns the bet history for a specific market.
      parameters:
        - in: path
          name: marketId
          required: true
          description: Numeric identifier of the market.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Bets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketBetsEnvelopeResponse'
        '400':
          description: Invalid market ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '404':
          description: Market not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                notFound:
                  value:
                    ok: false
                    reason: NOT_FOUND
        '500':
          description: Unexpected server error while fetching bets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    MiddlewareRateLimitExceeded:
      description: Request rejected by shared rate-limiting middleware before handler-specific logic ran.
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainTextErrorResponse'
  schemas:
...
    ReasonResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        reason:
          $ref: '#/components/schemas/FailureReason'
      required: [ok, reason]
    PlainTextErrorResponse:
      type: string
      description: Plain text error emitted by shared middleware before request-specific handlers run.
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    LoginResponse:
      type: object
      required: [ok, result]
      properties:
        ok:
          type: boolean
          enum: [true]
        result:
          $ref: '#/components/schemas/LoginResult'
...
    Economics:
      type: object
      properties:
        marketcreation:
          $ref: '#/components/schemas/MarketCreation'
        marketincentives:
          $ref: '#/components/schemas/MarketIncentives'
        user:
          $ref: '#/components/schemas/UserEconomics'
        betting:
          $ref: '#/components/schemas/Betting'
...
    FrontendConfig:
      type: object
      properties:
        charts:
          $ref: '#/components/schemas/FrontendCharts'
...